add Controllers

+++ src/main/java/pwr/piisw/cinema/application/service/CinemaService.java
@@ -1,19 +1,19 @@
 package pwr.piisw.cinema.application.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pwr.piisw.cinema.application.entity.Cinema;
 import pwr.piisw.cinema.application.repository.CinemaRepository;
+import pwr.piisw.cinema.exception.CinemaException;
+import pwr.piisw.cinema.exception.CinemaExceptionType;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CinemaService {
 
     private final CinemaRepository cinemaRepository;
 
-    @Autowired
     public CinemaService(CinemaRepository cinemaRepository) {
         this.cinemaRepository = cinemaRepository;
     }
@@ -22,19 +22,46 @@ public class CinemaService {
         return cinemaRepository.findAll();
     }
 
-    public Optional<Cinema> findById(Integer id) {
-        return cinemaRepository.findById(id);
+    public Cinema findById(Integer id) {
+        return cinemaRepository.findById(id)
+                .orElseThrow(() -> new CinemaException(CinemaExceptionType.CINEMA_NOT_FOUND));
     }
 
     public List<Cinema> findByCity(String city) {
         return cinemaRepository.findByCity(city);
     }
 
+    public List<Cinema> findCinemasByName(String name) {
+        return cinemaRepository.findCinemasByName(name);
+    }
+
+    public List<Cinema> searchCinemas(String keyword) {
+        return cinemaRepository.searchCinemas(keyword);
+    }
+
+    @Transactional
     public Cinema save(Cinema cinema) {
         return cinemaRepository.save(cinema);
     }
 
+    @Transactional
+    public void updateAddress(Integer cinemaId, String address) {
+        if (!cinemaRepository.existsById(cinemaId)) {
+            throw new CinemaException(CinemaExceptionType.CINEMA_NOT_FOUND);
+        }
+        cinemaRepository.updateAddress(cinemaId, address);
+    }
+
+    @Transactional
     public void deleteById(Integer id) {
+        if (!cinemaRepository.existsById(id)) {
+            throw new CinemaException(CinemaExceptionType.CINEMA_NOT_FOUND);
+        }
         cinemaRepository.deleteById(id);
+    }
+
+    @Transactional
+    public void deleteCinemasByCity(String city) {
+        cinemaRepository.deleteCinemasByCity(city);
     }
 }
